Move reading network parameters
Move reading of network parameters on startup to separate
NetworkParametersReader class. Add tests.
