+++ node/src/test/kotlin/net/corda/node/services/network/NetworkMapClientTest.kt
@@ -1,11 +1,21 @@
 package net.corda.node.services.network
 
+import com.google.common.jimfs.Configuration
+import com.google.common.jimfs.Jimfs
 import net.corda.core.crypto.Crypto
 import net.corda.core.crypto.SignedData
 import net.corda.core.crypto.sha256
 import net.corda.core.crypto.sign
+import net.corda.core.internal.*
+import net.corda.core.node.NetworkParameters
+import net.corda.core.serialization.deserialize
 import net.corda.core.serialization.serialize
 import net.corda.core.utilities.seconds
+import net.corda.node.internal.NetworkParametersReader
+import net.corda.nodeapi.internal.network.NETWORK_PARAMS_FILE_NAME
+import net.corda.nodeapi.internal.network.NETWORK_PARAMS_UPDATE_FILE_NAME
+import net.corda.nodeapi.internal.network.NetworkParametersCopier
+import net.corda.nodeapi.internal.network.verifiedNetworkMapCert
 import net.corda.testing.core.ALICE_NAME
 import net.corda.testing.core.BOB_NAME
 import net.corda.testing.core.DEV_ROOT_CA
@@ -24,9 +34,11 @@ import org.junit.Rule
 import org.junit.Test
 import java.io.IOException
 import java.net.URL
+import java.nio.file.Path
 import java.time.Instant
 import java.time.temporal.ChronoUnit
 import kotlin.test.assertEquals
+import kotlin.test.assertFalse
 
 class NetworkMapClientTest {
     @Rule
@@ -115,5 +127,20 @@ class NetworkMapClientTest {
         val keyPair = Crypto.generateKeyPair()
         val signedHash = SignedData(hash2.serialize(), keyPair.private.sign(hash2.serialize().bytes, keyPair.public))
         networkMapClient.ackNetworkParametersUpdate(signedHash)
+    }
+
+    @Test
+    fun `read correct set of parameters from file`() {
+        val fs = Jimfs.newFileSystem(Configuration.unix())
+        val baseDirectory = fs.getPath("/node").createDirectories()
+        val oldParameters = testNetworkParameters(emptyList(), epoch = 1)
+        NetworkParametersCopier(oldParameters).install(baseDirectory)
+        NetworkParametersCopier(server.networkParameters, update = true).install(baseDirectory) // Parameters update file.
+        val parameters = NetworkParametersReader(DEV_ROOT_CA.certificate, networkMapClient, baseDirectory).networkParameters
+        assertFalse((baseDirectory / NETWORK_PARAMS_UPDATE_FILE_NAME).exists())
+        assertEquals(server.networkParameters, parameters)
+        // Parameters from update should be moved to `network-parameters` file.
+        val parametersFromFile = (baseDirectory / NETWORK_PARAMS_FILE_NAME).readAll().deserialize<SignedDataWithCert<NetworkParameters>>().verifiedNetworkMapCert(DEV_ROOT_CA.certificate)
+        assertEquals(server.networkParameters, parametersFromFile)
     }
 }
